# Prevent memoize cache miss by passing "node.key" instead of "node"