This PR is an experiment to add a deprecation on methods using `Normalize.key` and use `node.key` instead of `node` when using `Node` methods.

The main benefits is to prevent cache misses in the memoization. Also since the memoization uses `Immutable.is` comparing string keys instead of record should be faster.

(I'll post a comment if I have interesting result from benchmarks).