@SamyPesse 2016-11-03
@Soreine @ianstormtaylor

Benchmarks logs:

#### [This branch](https://travis-ci.org/GitbookIO/slate/jobs/172799527)

```
Benchmarks
   delete-backward-deep
     Current:	 258 ops/sec (108 runs sampled)
     rme: ±3.59%
   delete-backward-normal
     Current:	 331 ops/sec (115 runs sampled)
     rme: ±3.11%
   delete-backward-wide
     Current:	 265 ops/sec (96 runs sampled)
     rme: ±7.82%
   first-render-normal
     Current:	 63.98 ops/sec (122 runs sampled)
     rme: ±3.46%
   insert-text-normal
     Current:	 760 ops/sec (93 runs sampled)
     rme: ±5.03%
   memoize-util
     Current:	 2,554 ops/sec (94 runs sampled)
     rme: ±6.92%
   normalize-document-deep
     Current:	 6,078 ops/sec (82 runs sampled)
     rme: ±8.47%
   normalize-document-normal
     Current:	 9,011 ops/sec (82 runs sampled)
     rme: ±4.01%
   normalize-document-wide
     Current:	 7,930 ops/sec (80 runs sampled)
     rme: ±9.03%
   render-split-block
     Current:	 35.12 ops/sec (113 runs sampled)
     rme: ±4.46%
   split-block-deep
     Current:	 270 ops/sec (105 runs sampled)
     rme: ±5.05%
   split-block-normal
     Current:	 434 ops/sec (104 runs sampled)
     rme: ±6.64%
   split-block-wide
     Current:	 331 ops/sec (95 runs sampled)
     rme: ±7.34%
```

#### [Origin](https://travis-ci.org/ianstormtaylor/slate/jobs/172684545)

```
Benchmarks
   delete-backward-deep
     Current:	 254 ops/sec (108 runs sampled)
     rme: ±2.87%
   delete-backward-normal
     Current:	 296 ops/sec (105 runs sampled)
     rme: ±5.38%
   delete-backward-wide
     Current:	 280 ops/sec (96 runs sampled)
     rme: ±5.78%
   first-render-normal
     Current:	 57.06 ops/sec (125 runs sampled)
     rme: ±4.20%
   insert-text-normal
     Current:	 561 ops/sec (90 runs sampled)
     rme: ±7.63%
   memoize-util
     Current:	 3,475 ops/sec (97 runs sampled)
     rme: ±5.80%
   normalize-document-deep
     Current:	 6,274 ops/sec (82 runs sampled)
     rme: ±7.35%
   normalize-document-normal
     Current:	 6,398 ops/sec (81 runs sampled)
     rme: ±8.93%
   normalize-document-wide
     Current:	 4,213 ops/sec (80 runs sampled)
     rme: ±9.17%
   render-split-block
     Current:	 25.61 ops/sec (121 runs sampled)
     rme: ±5.73%
   split-block-deep
     Current:	 163 ops/sec (110 runs sampled)
     rme: ±5.01%
   split-block-normal
     Current:	 233 ops/sec (101 runs sampled)
     rme: ±7.38%
   split-block-wide
     Current:	 192 ops/sec (94 runs sampled)
     rme: ±6.27%
```

You can see that basically almost everything looks faster ⚡️